--- a/.travis.yml
+++ b/.travis.yml
@@ -18,3 +18,4 @@ script:
   - 'if [[ $TRAVIS_GO_VERSION = 1.8* ]]; then ! go tool vet -all . 2>&1 | grep -vF .pb.go:; fi' # https://github.com/golang/protobuf/issues/214
   - make test testrace
   - 'if [[ $TRAVIS_GO_VERSION = 1.8* ]]; then staticcheck -ignore google.golang.org/grpc/transport/transport_test.go:SA2002 ./...; fi' # TODO(menghanl): fix these
+  - 'if [[ $TRAVIS_GO_VERSION = 1.8* ]]; then ./benchmark/compare/compare_travis.sh || true; fi'