--- a/benchmark/benchmark16_test.go
+++ b/benchmark/benchmark16_test.go
@@ -30,80 +30,80 @@ import (
 
 func BenchmarkClientStreamc1(b *testing.B) {
 	grpc.EnableTracing = true
-	runStream(b, 1, 1, 1, 0, 0, 0)
+	runStream(b, nil, 1, 1, 1, 0, 0, 0)
 }
 
 func BenchmarkClientStreamc8(b *testing.B) {
 	grpc.EnableTracing = true
-	runStream(b, 8, 1, 1, 0, 0, 0)
+	runStream(b, nil, 8, 1, 1, 0, 0, 0)
 }
 
 func BenchmarkClientStreamc64(b *testing.B) {
 	grpc.EnableTracing = true
-	runStream(b, 64, 1, 1, 0, 0, 0)
+	runStream(b, nil, 64, 1, 1, 0, 0, 0)
 }
 
 func BenchmarkClientStreamc512(b *testing.B) {
 	grpc.EnableTracing = true
-	runStream(b, 512, 1, 1, 0, 0, 0)
+	runStream(b, nil, 512, 1, 1, 0, 0, 0)
 }
 func BenchmarkClientUnaryc1(b *testing.B) {
 	grpc.EnableTracing = true
-	runUnary(b, 1, 1, 1, 0, 0, 0)
+	runUnary(b, nil, 1, 1, 1, 0, 0, 0)
 }
 
 func BenchmarkClientUnaryc8(b *testing.B) {
 	grpc.EnableTracing = true
-	runUnary(b, 8, 1, 1, 0, 0, 0)
+	runUnary(b, nil, 8, 1, 1, 0, 0, 0)
 }
 
 func BenchmarkClientUnaryc64(b *testing.B) {
 	grpc.EnableTracing = true
-	runUnary(b, 64, 1, 1, 0, 0, 0)
+	runUnary(b, nil, 64, 1, 1, 0, 0, 0)
 }
 
 func BenchmarkClientUnaryc512(b *testing.B) {
 	grpc.EnableTracing = true
-	runUnary(b, 512, 1, 1, 0, 0, 0)
+	runUnary(b, nil, 512, 1, 1, 0, 0, 0)
 }
 
 func BenchmarkClientStreamNoTracec1(b *testing.B) {
 	grpc.EnableTracing = false
-	runStream(b, 1, 1, 1, 0, 0, 0)
+	runStream(b, nil, 1, 1, 1, 0, 0, 0)
 }
 
 func BenchmarkClientStreamNoTracec8(b *testing.B) {
 	grpc.EnableTracing = false
-	runStream(b, 8, 1, 1, 0, 0, 0)
+	runStream(b, nil, 8, 1, 1, 0, 0, 0)
 }
 
 func BenchmarkClientStreamNoTracec64(b *testing.B) {
 	grpc.EnableTracing = false
-	runStream(b, 64, 1, 1, 0, 0, 0)
+	runStream(b, nil, 64, 1, 1, 0, 0, 0)
 }
 
 func BenchmarkClientStreamNoTracec512(b *testing.B) {
 	grpc.EnableTracing = false
-	runStream(b, 512, 1, 1, 0, 0, 0)
+	runStream(b, nil, 512, 1, 1, 0, 0, 0)
 }
 func BenchmarkClientUnaryNoTracec1(b *testing.B) {
 	grpc.EnableTracing = false
-	runUnary(b, 1, 1, 1, 0, 0, 0)
+	runUnary(b, nil, 1, 1, 1, 0, 0, 0)
 }
 
 func BenchmarkClientUnaryNoTracec8(b *testing.B) {
 	grpc.EnableTracing = false
-	runUnary(b, 8, 1, 1, 0, 0, 0)
+	runUnary(b, nil, 8, 1, 1, 0, 0, 0)
 }
 
 func BenchmarkClientUnaryNoTracec64(b *testing.B) {
 	grpc.EnableTracing = false
-	runUnary(b, 64, 1, 1, 0, 0, 0)
+	runUnary(b, nil, 64, 1, 1, 0, 0, 0)
 }
 
 func BenchmarkClientUnaryNoTracec512(b *testing.B) {
 	grpc.EnableTracing = false
-	runUnary(b, 512, 1, 1, 0, 0, 0)
+	runUnary(b, nil, 512, 1, 1, 0, 0, 0)
 }
 
 func TestMain(m *testing.M) {