--- a/benchmark/benchmark17_test.go
+++ b/benchmark/benchmark17_test.go
@@ -22,7 +22,6 @@ package benchmark
 
 import (
 	"fmt"
-	"os"
 	"testing"
 	"time"
 
@@ -32,13 +31,15 @@ import (
 
 func BenchmarkClient(b *testing.B) {
 	maxConcurrentCalls := []int{1, 8, 64, 512}
-	reqSizeBytes := []int{1, 1024, 1024 * 1024}
-	reqspSizeBytes := []int{1, 1024, 1024 * 1024}
+	reqSizeBytes := []int{1, 1024 * 1024}
+	reqspSizeBytes := []int{1, 1024 * 1024}
 	kbps := []int{0, 10240} // if non-positive, infinite
 	MTU := []int{0, 10}     // if non-positive, infinite
 	// When set the latency to 0 (no delay), the result is slower than the real result with no delay
 	// because latency simulation section has extra operations
 	latency := []time.Duration{0, 40 * time.Millisecond} // if non-positive, no delay.
+	s := stats.NewStats(38)
+	isMatched := false
 
 	for _, enableTracing := range []bool{true, false} {
 		grpc.EnableTracing = enableTracing
@@ -55,13 +56,23 @@ func BenchmarkClient(b *testing.B) {
 								b.Run(fmt.Sprintf("Unary-%s-kbps_%#v-MTU_%#v-maxConcurrentCalls_"+
 									"%#v-reqSize_%#vB-respSize_%#vB-latency_%s",
 									tracing, k, mtu, maxC, reqS, respS, ltc.String()), func(b *testing.B) {
-									runUnary(b, maxC, reqS, respS, k, mtu, ltc)
+									runUnary(b, s, maxC, reqS, respS, k, mtu, ltc)
+									isMatched = true
 								})
+								if isMatched {
+									fmt.Print(s.String())
+									isMatched = false
+								}
 								b.Run(fmt.Sprintf("Stream-%s-kbps_%#v-MTU_%#v-maxConcurrentCalls_"+
 									"%#v-reqSize_%#vB-respSize_%#vB-latency_%s",
 									tracing, k, mtu, maxC, reqS, respS, ltc.String()), func(b *testing.B) {
-									runStream(b, maxC, reqS, respS, k, mtu, ltc)
+									runStream(b, s, maxC, reqS, respS, k, mtu, ltc)
+									isMatched = true
 								})
+								if isMatched {
+									fmt.Print(s.String())
+									isMatched = false
+								}
 							}
 						}
 					}
@@ -70,8 +81,4 @@ func BenchmarkClient(b *testing.B) {
 		}
 	}
 
-}
-
-func TestMain(m *testing.M) {
-	os.Exit(stats.RunTestMain(m))
 }