1. Adding a file compare/main.go to compare the result of 2 benchmarks, thus the comparison can run locally.

2. Comparison in travis only runs in Go 1.8 version. You can open the travis of this pr. At the end of log, it prints result like
```
BenchmarkClient/Unary-Tracing-kbps_0-MTU_0-maxConcurrentCalls_1-reqSize_1048576B-respSize_1048576B-latency_40ms-2,     
Title       	Before         After   	Percentage 
operations           100           100            
     ns/op     239422160     239915920     0.21%  
      B/op      14594751      14299738    -2.02%  
 allocs/op          1936          1940     0.21%  
       50%    251.7915ms    251.9019ms     0.04%  
       90%    253.6508ms    255.2256ms     0.62%  
       99%    256.9438ms    259.5959ms     1.03%  
      100%    257.8455ms    261.3009ms     1.34%  
   Average    252.0007ms    252.4980ms     0.20%  
```